[MSE]During seek to unbuffered range before a buffered one, reenqueueing happens and playback becomes choppy
In reenqueueing function, we insert the new buffered samples following the seek and we also
insert the samples previously buffered before the seek. We'll remove them later, when the
new samples would ovelap the samples already buffered before the seek.

This commit insert only the new buffered samples following the seek.

Fix for #271

Signed-off-by: eocanha@igalia.com